+++ packages/react-property-selectors/src/__tests__/amount/amount-input-box-test-component.tsx
